+++ src/main/java/cart/repository/mapper/PointMapper.java
@@ -2,10 +2,11 @@ package cart.repository.mapper;
 
 import cart.domain.point.Point;
 import cart.entity.PointEntity;
+import java.math.BigDecimal;
 
 public class PointMapper {
 
     public static Point toPoint(final PointEntity pointEntity) {
-        return new Point(pointEntity.getPoint());
+        return new Point(BigDecimal.valueOf(pointEntity.getPoint()));
     }
 }
